Rename package back to 'kernel' from 'kernel-rt'; set buildid to .xenomai
'kernel-rt' is the name of the MRG RT_PREEMPT kernel.

This package is really a regular (although modified) kernel package
that builds a 'xenomai' flavor.

Therefore, it makes sense to set buildid to .xenomai to distinguish it
from another upstream kernel, whose name might conflict.  The
'vanilla' package with Xenomai and i-pipe disabled, called 'kernel' as
usual, will show the buildid in `uname -r`:  3.5.7-4.xenomai.el6.x86_64

The xenomai flavor package, called 'kernel-xenomai', will be called
'kernel-xenomai'.  The string 'xenomai' will be redundant in `uname
-r`: 3.5.7-4.xenomai.el6.xenomai.x86_64.  The first '.xenomai'
indicates the buildid.  The second '.xenomai' indicates the flavor.
With this scheme, the kernels may coexist.

Finally, this leaves room for a combo package.  The buildid would be
called something like 'ipipe', and then both 'xenomai' and 'rtai'
flavors could be built from the same source.